+++ apx_sdk/CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Added `NonCanonicalUri` type to `url::canonical` module.
 - Accept 'ap' URIs with `ap+ef61` scheme.
+- Added `Multibase` type to `multibase` module.
 
 ### Changed